Move store and reducer to central redux folder
