+++ src/redux/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from "redux";
-import auth from '../login/login.reducers'
+import auth from '../../components/login/login.reducers'
 import { routerReducer } from 'react-router-redux';
 
 export default combineReducers({ auth, router: routerReducer });
